bau: organise resources
Split into shared/idp/eidas

Pair: WillPa
